saturn_bot/requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
boto3>=1.34.0
prometheus-client>=0.20.0
docker>=7.0.0
requests>=2.31.0
aiohttp>=3.9.0
psutil>=5.9.0 